requirements.txt
kafka-python
pandas
psycopg2-binary
xgboost
scikit-learn